Test release

+++ src/main/java/agh/pin/pals/server/repositories/UserRepository.java
@@ -5,12 +5,14 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
 
     boolean existsByEmail(String email);
     boolean existsByUsername(String username);
     User findUserByEmail(String email);
     User findUserByUsername(String username);
+
+
 }
